chore: some clean up has been done, there are functions soon to be obsolete, they are for testing purposes at this moment
