+++ src/main.c
@@ -13,17 +13,15 @@
 #include "../inc/cubd.h"
 
 /*~~ main
-1. make_array should be obsolete once data->map is filled, its for testing now 
-2. initilize minimap is for the bonuse but utalized right now for visualizing 
-map creation, player movement and ray casting for now , it allso collects p_x and p_y
-untill parsing has handled that also
-3. mlx_key_hook() is for player movement and esc key function 
-4. mlx_loop(data->mlx) needs to be located where it is 
-5. mlx_terminate seg faults at the moment , could be an issue of needing a flag
+1. initilize minimap is for the bonuse but utalized right now for visualizing 
+map creation, player movement and ray casting for now
+2. mlx_key_hook() is for player movement and esc key function 
+3. mlx_loop(data->mlx) needs to be located where it is 
+4. mlx_terminate seg faults at the moment , could be an issue of needing a flag
 ~~*/
 int	main(int argc, char **argv)
 {
-	static t_data	*data; //~~ should it be static?
+	static t_data	*data;
 
 	if (argc == 2)
 	{
@@ -36,30 +34,29 @@ int	main(int argc, char **argv)
 			free_data(data);
 			return (FAILURE);
 		}
-		// printf("bug hunting 1\n");
 		printer(data);
 		if (open_window(data) == FAILURE)
 		{
 			printf("error requires freeing in main \n");
 			return (FAILURE);
 		}
-		//collect_ray(data); in minimap init for now
-		// initilize_other?
-		if (initlize_minimap(data) == 1)
+		// collect_ray(data); in minimap init for now
+		// **init_3d(data);
+		if (initlize_minimap(data) == FAILURE)
 		{
 			printf("error requires freeing in main\n");
-			return (-1);
+			return (FAILURE);
 		}
 		mlx_key_hook(data->mlx, &keyhookfunc, data);
-		
 		//~~ bonus animation if wanted needs to start here
 		// mlx_loop_hook(data->mlx, &animation_fucn, &data);
-		//printf("bug huntinig\n");
 		mlx_loop(data->mlx);
 		printer(data);
-		//free_array(data->file);
-		
-		//mlx_terminate(data->mlx);
+		// if (flag == true)
+		{
+			// free_array(data->file);
+			// mlx_terminate(data->mlx);
+		}
 	}
 	else
 		usage();
