+++ src/test.cub
@@ -3,6 +3,6 @@
 1000000000000000000000001
 1000000000000000000000001
 1000000000000000000000001
-1000000000000000000000001
-100000000000N000000000001
+1000000000000000000001111
+100000000000N000000001000
 1111111111111111111111111
